Wander
Added wander and changed some code in add new Random unit. Wander has arrive functionality right now. Not sure if I need to implement Wander as a child class of Face similar to the reading. Also not sure if I should do this with ArriveAndFace and WanderAndChase. Also not sure how to implement private variables without making the constructor parameters long and gross.