Wander And WanderAndChase
Wander steerings are implemented. Not using face steering to face towards the point, takes to long to rotate towards so it always faces straight. Face Steering has a issue where it will continue to spin. Think has an issue with getting the range of the radian to become less than pi and negative pi.